pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opendem"
version = "0.1.0"
description = "A bootstrapper library that loads library referenced in yml and runs library on that yml file"
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name="Stefan Haselwimmer", email="haselwimmer@gmail.com" }
]
dependencies = [
    "pyyaml",
    "requests"
]

[project.scripts]
openlibrary = "openlibrary.core:main"

[tool.setuptools.packages.find]
where = ["."]